Add type-cast, making sure that the unary minus is handled correctly on any compiler

+++ generic/tclUtil.c
@@ -3335,7 +3335,7 @@ TclFormatInt(buffer, n)
      * Generate the characters of the result backwards in the buffer.
      */
 
-    intVal = (n < 0 ? -n : n);
+    intVal = (n < 0 ? -(unsigned long)n : (unsigned long)n);
     i = 0;
     buffer[0] = '\0';
     do {
